[HUDI-3959] Rename spark read client to avoid confused meaning
